test: api server errors
